grouping with grading model
